+++ bpdm-common-test/src/main/kotlin/org/eclipse/tractusx/bpdm/test/testdata/gate/v7/GateAssertRepositoryV7.kt
@@ -110,6 +110,7 @@ class GateAssertRepositoryV7(
             "site.${SiteRepresentationOutputDto::siteBpn.name}",
             "site.${SiteRepresentationOutputDto::confidenceCriteria.name}",
             "site.${SiteRepresentationOutputDto::states.name}",
+            "site.${SiteRepresentationOutputDto::goldenRecordRelations.name}",
             "site.${SiteRepresentationOutputDto::updatedAt.name}",
             // address: keep only name / addressType / physical & alternative postal address
             "address.${AddressComponentOutputDto::addressBpn.name}",
@@ -385,7 +386,10 @@ class GateAssertRepositoryV7(
         )
 
     private fun SiteRepresentationOutputDto.sortContent() =
-        copy(states = states.sortedBy { it.validFrom?.toString() })
+        copy(
+            states = states.sortedBy { it.validFrom?.toString() },
+            goldenRecordRelations = goldenRecordRelations.sortedBy { it.sourceBpn }
+        )
 
     private fun AddressComponentOutputDto.sortContent() =
         copy(