+++ bpdm-orchestrator/src/main/resources/db/migration/V7_5_0_2__add_site_golden_record_relations.sql
@@ -1,0 +1,10 @@
+CREATE TABLE business_partner_site_golden_record_relations
+(
+    task_id       BIGINT       NOT NULL,
+    relation_type VARCHAR(255) NOT NULL,
+    source_bpn    VARCHAR(255) NOT NULL,
+    target_bpn    VARCHAR(255) NOT NULL,
+    CONSTRAINT fk_site_golden_record_relations_tasks FOREIGN KEY (task_id) REFERENCES golden_record_tasks (id)
+);
+
+CREATE INDEX index_site_golden_record_relations_task_id ON business_partner_site_golden_record_relations (task_id);